// frontend/src/app/models/intern/intern.module.ts
export interface Intern {
  id: string;
  name: string;
  email: string;
  department: string;
  mentor: string;
  university: string;
  status: string;
  startDate: string;
  endDate: string;
}

export interface InternFormData {
  name: string;
  email: string;
  department: string;
  mentor: string;
  university: string;
  status: string;
  startDate: string;
  endDate: string;
}